Split visibility tests into their own file
